Update_更新内核.sh

+++ clash/scripts/更新内核.sh
@@ -1,15 +1,22 @@
-url=$(curl -s https://api.github.com/repos/MetaCubeX/Clash.Meta/releases/tags/Prerelease-Alpha \
-| grep "download/Prerelease-Alpha/Clash.Meta-android-arm64-alpha" \
+tag="latest"
+#tag="tags/Prerelease-Alpha"
+
+url=$(curl -s https://api.github.com/repos/MetaCubeX/Clash.Meta/releases/${tag} \
+| grep "android-arm64" \
 | cut -d : -f 2,3 \
 | tr -d \")
 
-curl -L $url -o clash.gz
+url=${url#*https}
 
-gunzip clash.gz
+curl -L "https${url}" -o clash.gz
 
-mv /data/adb/modules/Clash_For_Magisk/system/bin/clash /data/adb/modules/Clash_For_Magisk/system/bin/clash.bak
-
-mv clash /data/adb/modules/Clash_For_Magisk/system/bin/clash
-
-chmod 777 /data/adb/modules/Clash_For_Magisk/system/bin/clash
-echo 已安装最新版 meta alpha内核
+if [ -e clash.gz ]; then
+    gunzip clash.gz
+    
+    mv /data/adb/modules/Clash_For_Magisk/system/bin/clash /data/adb/modules/Clash_For_Magisk/system/bin/clash.bak
+    mv clash /data/adb/modules/Clash_For_Magisk/system/bin/clash
+    chmod 777 /data/adb/modules/Clash_For_Magisk/system/bin/clash
+    
+    vvv=$(/data/adb/modules/Clash_For_Magisk/system/bin/clash -v | cut -d ' ' -f 2,3)
+    echo "已安装最新版$vvv内核"
+fi
